+++ lib/src/framework/core/dependencies/global_widgets/hover_global_widgets.dart
@@ -15,5 +15,6 @@ class HoverGlobalWidgets {
 }
 
 class HoverGlobalWidgetsProvider extends Provider<HoverGlobalWidgets> {
-  HoverGlobalWidgetsProvider(HoverGlobalWidgets globalElements) : super(create: (_) => globalElements);
+  HoverGlobalWidgetsProvider(HoverGlobalWidgets globalElements)
+      : super(create: (_) => globalElements);
 }