+++ lib/src/framework/helpers/hover_snackbar_helper.dart
@@ -1,13 +1,15 @@
 import 'package:flutter/material.dart';
 
 class HoverSnackBarHelper {
-  static const Duration defaultSnackBarDuration = Duration(seconds: 2);
+  static const Duration defaultSnackBarDuration = const Duration(seconds: 2);
 
-  static void showSnackBar(BuildContext context, Widget content, {Duration duration: defaultSnackBarDuration}) {
-    Scaffold.of(context).showSnackBar(SnackBar(content: content, duration: duration));
+  static void showSnackBar(BuildContext context, SnackBar snackBar) {
+    Scaffold.of(context).showSnackBar(snackBar);
   }
 
-  static void showPlainSnackBar(BuildContext context, String message, {Duration duration: defaultSnackBarDuration}) {
-    showSnackBar(context, Text(message), duration: duration);
+  static void showPlainSnackBar(BuildContext context, String message,
+      {Duration duration: defaultSnackBarDuration}) {
+    Scaffold.of(context)
+        .showSnackBar(SnackBar(content: Text(message), duration: duration));
   }
 }