+++ lib/src/widgets/hover_widgets.dart
@@ -15,6 +15,7 @@ export 'custom/hover_square_image.dart';
 export 'custom/hover_image_slider.dart';
 export 'custom/hover_search_bar.dart';
 export 'custom/hover_base_card.dart';
+export 'custom/hover_snack_bar.dart';
 
 //Typography
 export 'typography/typography.dart';